fix(pipeline): replace false-positive v4 certification

+++ scripts/pipeline_steps/_runner.py
@@ -335,6 +335,7 @@ def run_scripts(
     finally:
         if temp_db_path:
             try:
-                os.unlink(temp_db_path)
+                if os.environ.get("BET_PIPELINE_OFFLINE_TEST_MODE") != "1":
+                    os.unlink(temp_db_path)
             except OSError:
                 pass
